Fixed profile queries

+++ app/src/main/java/it/polito/timebanking/repository/User.kt
@@ -7,12 +7,13 @@ import androidx.room.PrimaryKey
 @Entity(tableName = "users", indices = [Index("fullname","nickname","email","location")])
 class User {
     @PrimaryKey(autoGenerate = true)
-    var id:Int = 0
+    var id:Long = 0
     //Lab2
     var fullname:String = ""
     var nickname:String = ""
     var email:String = ""
     var location:String = ""
+
     //Lab3
     //var slots:List<Int>???
 }
