+++ app/src/main/java/it/polito/timebanking/repository/UserDao.kt
@@ -1,10 +1,7 @@
 package it.polito.timebanking.repository
 
 import androidx.lifecycle.LiveData
-import androidx.room.Dao
-import androidx.room.Insert
-import androidx.room.Query
-import androidx.room.Update
+import androidx.room.*
 
 @Dao
 interface UserDao {
@@ -12,7 +9,7 @@ interface UserDao {
     fun findAll(): LiveData<List<User>>
 
     @Query("SELECT * from users WHERE id= :id")
-    fun searchUserByID(id:Int): LiveData<User>
+    fun searchUserByID(id:Long): LiveData<User>
 
     //If you insert one entity at a time, you can get its ID back as a Long
     @Insert
